perl5: Use newer compiler for perl_tsa_mutex_lock
Use a newer compiler (Apple clang 6.1 or later or open source clang 3.6
or later) to ensure that perl_tsa_mutex_lock gets compiled in, which is
needed if any module would ever get compiled with such a newer compiler.

See: https://github.com/Perl/perl5/issues/18991

Perl is only in the clang dependency chain if clang's +analyzer variant
is selected, which it no longer is by default on 10.6 or earlier. For
clang 3.3 this happened in 9a0f6dd; for clang 3.4 in 640b7e0; for clang
3.7 in 9809d4b; for clang 5.0, 6.0, 7.0, 8.0, 9.0, and 10 in 6b338b8;
and for clang 11, 12, and devel in 2648d97.

Closes: https://trac.macports.org/ticket/57648